% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLD.R
\name{sim_plate_count}
\alias{sim_plate_count}
\title{Simulate plating}
\usage{
sim_plate_count(n_re, n_wt, n_plates, n_sample)
}
\arguments{
\item{n_re}{Number of resistant cells in the tube}

\item{n_wt}{Number of wild type cells in the tube}

\item{n_plates}{Number of replicate plates per experiment}

\item{n_sample}{Number of cells to plate from a tube}
}
\value{
A vector of integers for numbers of colonies on plates with phages
}
\description{
Simulate plating cell culture from a tube containing phage-resistant and wild type (phage-sensitive) cells to plates with phages
}
