% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLD.R
\name{sim_tube_count}
\alias{sim_tube_count}
\title{Simulate growth}
\usage{
sim_tube_count(n_gens, mut_rate, ncells_init, ncells_res_init = 0)
}
\arguments{
\item{n_gens}{Number of generations in tube}

\item{mut_rate}{Mutation rate}

\item{ncells_init}{Initial number of cells in a tube transferred from a flask before growth}

\item{ncells_res_init}{Number of mutant cells in a tube before growth representing standing variation}
}
\value{
A vector of integers for numbers of colonies on plates with phages
}
\description{
Simulate cell growth in a tube
}
