% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLD.R
\name{simLD_ind}
\alias{simLD_ind}
\title{Simulate induced mutation model}
\usage{
simLD_ind(n_plates, n_sample, mut_rate)
}
\arguments{
\item{n_plates}{Number of replicate plates per experiment.}

\item{n_sample}{Number of cells to plate from a tube after growth}

\item{mut_rate}{Mutation rate}
}
\value{
A named vector of four: mean number of resistant colonies in experiment A (plating on n_plates plates from one tube), mean number of resistant colonies in experiment B (plating on n_plates plates from n_plates tubes), variance in experiment A, and variance in experiment B.
}
\description{
Simulate a Luria-Delbruck experiment assuming induced mutation by exposure to phages
}
