% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLD.R
\name{simLD_spo}
\alias{simLD_spo}
\title{Simulate spontaneous mutation model}
\usage{
simLD_spo(
  n_gens,
  mut_rate,
  ncells_init,
  n_sample,
  n_plates,
  ncells_res_init = 0
)
}
\arguments{
\item{n_gens}{Number of generations in tube}

\item{mut_rate}{Mutation rate}

\item{ncells_init}{Initial number of cells in a tube transferred from a flask before growth}

\item{n_sample}{Number of cells to plate from a tube after growth}

\item{n_plates}{Number of replicate plates per experiment.}

\item{ncells_res_init}{Number of mutant cells in a tube before growth representing standing variation}
}
\value{
A named vector of four: mean number of resistant colonies in experiment A (plating on n_plates plates from one tube), mean number of resistant colonies in experiment B (plating on n_plates plates from n_plates tubes), variance in experiment A, and variance in experiment B.
}
\description{
Simulate a Luria-Delbruck experiment assuming spontaneous mutation of resistance to phages
}
